backend/pyproject.toml
[tool.poetry]
name = "contest-platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "*"
sphinx-needs = "*"
"sphinxcontrib.plantuml" = "*"
djangorestframework = "*"
django-cors-headers = "*"
psycopg = "*"
pre-commit = "*"
python-dotenv = "*"
pipfile = "*"
drf-spectacular = "*"
drf-spectacular-sidecar = "*"
selenium = "*"
pytest = "*"
sentry-sdk = "*"
django-dramatiq = "^0.11.6"
requests = "^2.31.0"
dramatiq = {extras = ["rabbitmq", "watch"], version = "^1.16.0"}
WeasyPrint = "*"
black = "^24.4.2"


[tool.poetry.group.dev.dependencies]
sphinx = "*"
"sphinxcontrib.httpdomain" = "*"
sphinxcontrib-mermaid = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
